Regresssion model with Keras
First version of the regression model with Keras to model data and evaluate the performance of the model when changing data, number of epochs and number of hidden layers. .
A is the base line.
B normalizes the data.
C increases the number of epochs.
D increases the number of hidden layers.